refactor: extract capability::state (Capabilities/EmbedderState/ServerLimits)
Pure lift-and-shift, first step of splitting the capability.rs
monolith into a capability/ module directory (matching the config/ and
storage/memory/ split convention). No logic or behavior change:
EmbedderState, ServerLimits, and Capabilities (plus their tests) move
into crates/spelunk-cli/src/capability/state.rs unchanged; capability.rs
re-exports them so every existing caller keeps working.
